+++ hash.h
@@ -9,11 +9,12 @@
 #define hash_h
 
 #include <stdbool.h>
+
 #include "datum.h"
 #include "mmgt.h"
 #include "list.h"
 
-typedef int (*scf_hash_func)(scf_datum key);
+typedef size_t (*scf_hash_func)(scf_datum key);
 typedef bool (*scf_comparison_func)(scf_datum k1, scf_datum k2);
 
 typedef struct {
@@ -24,8 +25,8 @@ typedef struct {
 typedef struct {
     scf_hash_func hash_func;
     scf_comparison_func comparison_func;
-    int size;
-    int capacity;
+    size_t size;
+    size_t capacity;
     int max_collisions;
     scf_dictionary_item *items;
 } scf_dictionary;
@@ -34,7 +35,7 @@ typedef struct {
     scf_dictionary dictionary;
 } scf_set;
 
-scf_dictionary scf_dictionary_create(scf_operation *operation, scf_hash_func, scf_comparison_func, int initial_capacity);
+scf_dictionary scf_dictionary_create(scf_operation *operation, scf_hash_func, scf_comparison_func, size_t initial_capacity);
 
 scf_datum scf_dictionary_add(scf_dictionary *, scf_datum key, scf_datum value);
 
@@ -44,7 +45,7 @@ scf_datum *scf_dictionary_lookup(const scf_dictionary *, scf_datum key);
 
 scf_list scf_dictionary_get_items(scf_operation *operation, const scf_dictionary *);
 
-scf_set scf_set_create(scf_operation *operation, scf_hash_func, scf_comparison_func, int initial_capacity);
+scf_set scf_set_create(scf_operation *operation, scf_hash_func, scf_comparison_func, size_t initial_capacity);
 
 bool scf_set_add(scf_set *s, scf_datum item);
